+++ repos/common/tests/src/commonMain/kotlin/CommonRepoTests.kt
@@ -1,0 +1,6 @@
+package dev.inmo.micro_utils.repos.common.tests
+
+abstract class CommonRepoTests<T> {
+    protected open val testSequencesSize = 1000
+    protected abstract val repoCreator: suspend () -> T
+}